+++ packages/core/table-tree/examples/slow-load.js
@@ -11,12 +11,18 @@ export default class SlowLoad extends Component<
     tableData: null,
   };
 
+  dataTimeoutId = undefined;
+
   componentDidMount() {
-    setTimeout(() => {
+    this.dataTimeoutId = setTimeout(() => {
       this.setState({
         tableData: staticData.children,
       });
     }, 3000);
+  }
+
+  componentWillUnmount() {
+    clearTimeout(this.dataTimeoutId);
   }
 
   render() {