+++ packages/editor/renderer/src/__tests__/visual-regression/overflow.ts
@@ -23,5 +23,6 @@ describe('Snapshot Test: Overflow shadows', () => {
 
   it(`should render right shadows`, async () => {
     await initRenderer(page, document);
+    await page.waitFor('.code-block');
   });
 });