+++ packages/navigation/app-navigation/src/components/AppNavigation/index.tsx
@@ -3,8 +3,8 @@ import { NavigationAnalyticsContext } from '@atlaskit/analytics-namespaced-conte
 import { jsx } from '@emotion/core';
 
 import { AppNavigationTheme, ThemeProvider, defaultTheme } from '../../theme';
-import { ThemedPrimaryButton } from '../PrimaryButton';
 import { containerCSS, leftCSS, rightCSS } from './styles';
+import { PrimaryItemsContainer } from '../PrimaryItemsContainer';
 import { AppNavigationProps } from './types';
 
 const analyticsData = {
@@ -20,23 +20,23 @@ export const AppNavigation = (
     renderAppSwitcher: AppSwitcher,
     renderCreate: Create,
     renderHelp: Help,
+    renderNotifications: Notifications,
     renderProductHome: ProductHome,
     renderProfile: Profile,
-    renderNotifications: Notifications,
     renderSearch: Search,
+    renderSignIn: SignIn,
     renderSettings: Settings,
+    moreLabel,
     theme,
   } = props;
 
   return (
-    <ThemeProvider theme={theme}>
+    <ThemeProvider value={theme}>
       <NavigationAnalyticsContext data={analyticsData}>
-        <div css={containerCSS}>
+        <div css={containerCSS(theme)}>
           <div css={leftCSS}>
             {ProductHome && <ProductHome />}
-            {primaryItems.map(props => (
-              <ThemedPrimaryButton key={props.id} {...props} />
-            ))}
+            <PrimaryItemsContainer moreLabel={moreLabel} items={primaryItems} />
           </div>
           <div css={rightCSS}>
             {Create && <Create />}
@@ -45,7 +45,8 @@ export const AppNavigation = (
             {Notifications && <Notifications />}
             {Settings && <Settings />}
             {Help && <Help />}
-            <Profile />
+            {SignIn && <SignIn />}
+            {Profile && <Profile />}
           </div>
         </div>
       </NavigationAnalyticsContext>
@@ -55,5 +56,6 @@ export const AppNavigation = (
 
 AppNavigation.defaultProps = {
   primaryItems: [],
+  moreLabel: '…',
   theme: defaultTheme,
 };