+++ packages/navigation/app-navigation/src/components/AppSwitcher/index.tsx
@@ -1,22 +1,17 @@
 import React from 'react';
 import AppSwitcherIcon from '@atlaskit/icon/glyph/app-switcher';
 
-import { ThemedIconButton } from '../IconButton';
-import { TriggerManager } from '../TriggerManager';
+import { IconButton } from '../IconButton';
 import { AppSwitcherProps } from './types';
 
 export const AppSwitcher = (props: AppSwitcherProps) => {
-  const { tooltip, ...triggerManagerProps } = props;
+  const { tooltip, ...iconButtonProps } = props;
 
   return (
-    <TriggerManager {...triggerManagerProps}>
-      {({ onTriggerClick }) => (
-        <ThemedIconButton
-          icon={<AppSwitcherIcon label={tooltip} />}
-          onClick={onTriggerClick}
-          tooltip={tooltip}
-        />
-      )}
-    </TriggerManager>
+    <IconButton
+      icon={<AppSwitcherIcon label={tooltip} />}
+      tooltip={tooltip}
+      {...iconButtonProps}
+    />
   );
 };