+++ packages/navigation/atlassian-switcher/docs/04-non-react-apps.tsx
@@ -1,5 +1,0 @@
-import { md } from '@atlaskit/docs';
-
-export default md`
-  To render the Atlassian Switcher on non react apps, check it out the [Atlassian Switcher Vanilla](/packages/navigation/atlassian-switcher-vanilla) wrapper.
-`;