+++ packages/search/global-search/src/components/SearchResultsUtil.tsx
@@ -27,11 +27,20 @@ export function getConfluenceAdvancedSearchLink(query?: string) {
   return `/wiki/dosearchsite.action${queryString}`;
 }
 
-export function getJiraAdvancedSearchUrl(
-  entityType: JiraEntityTypes,
-  query?: string,
-  enableIssueKeySmartMode?: boolean,
-) {
+type Props = {
+  entityType: JiraEntityTypes;
+  query?: string;
+  enableIssueKeySmartMode?: boolean;
+  isJiraPeopleProfilesEnabled?: boolean;
+};
+
+export function getJiraAdvancedSearchUrl(props: Props) {
+  const {
+    entityType,
+    query,
+    enableIssueKeySmartMode,
+    isJiraPeopleProfilesEnabled,
+  } = props;
   switch (entityType) {
     case JiraEntityTypes.Issues:
       return !enableIssueKeySmartMode && query && +query
@@ -44,7 +53,9 @@ export function getJiraAdvancedSearchUrl(
     case JiraEntityTypes.Projects:
       return `/projects?contains=${query}`;
     case JiraEntityTypes.People:
-      return `/people/search?q=${query}`;
+      return !!isJiraPeopleProfilesEnabled
+        ? `/jira/people/search?q=${query}`
+        : `/people/search?q=${query}`;
   }
 }
 
@@ -57,7 +68,13 @@ export function redirectToJiraAdvancedSearch(
   entityType: JiraEntityTypes,
   query = '',
 ) {
-  window.location.assign(getJiraAdvancedSearchUrl(entityType, query, true));
+  window.location.assign(
+    getJiraAdvancedSearchUrl({
+      entityType,
+      query,
+      enableIssueKeySmartMode: true,
+    }),
+  );
 }
 
 export function take<T>(array: Array<T>, n: number) {