+++ packages/core/popup/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @atlaskit/popup
 
+## 0.1.2
+
+### Patch Changes
+
+- [patch][242dd7a06d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/242dd7a06d):
+
+  Expose additional types
+
+## 0.1.1
+
+### Patch Changes
+
+- [patch][583a9873ef](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/583a9873ef):
+
+  Provided better description for popup types
+
 ## 0.1.0
 
 ### Minor Changes